[~] Pulled 1.20.1 port from ihatecsv's fork.

+++ common/src/main/java/com/starfish_studios/naturalist/common/entity/core/ai/navigation/BetterWallClimberNavigation.java
@@ -50,7 +50,7 @@ public class BetterWallClimberNavigation extends GroundPathNavigation {
                 if (!this.pathToPosition.closerToCenterThan(this.mob.position(),
                         Math.max(this.mob.getBbWidth(), 1.0D))
                         && (!(this.mob.getY() > (double) this.pathToPosition.getY())
-                        || !(new BlockPos(this.pathToPosition.getX(), this.mob.getY(),
+                        || !(new BlockPos(this.pathToPosition.getX(), (int) this.mob.getY(),
                         this.pathToPosition.getZ())).closerToCenterThan(this.mob.position(),
                         Math.max(this.mob.getBbWidth(), 1.0D)))) {
                     this.mob.getMoveControl().setWantedPosition(this.pathToPosition.getX(), this.pathToPosition.getY(),
